Rebase on GEODE develop (#7372)

+++ dev-tools/release/license_review.sh
@@ -132,6 +132,7 @@ function listJarsInWar() {
 }
 
 function extractLicense() {
+  [ "$SKIP_LICENSES" != "true" ] || return 0
   war=$1
   rm -Rf tmpl
   mkdir tmpl
@@ -227,7 +228,7 @@ log4j-
 lucene-
 mapstruct
 micrometer-core
-netty-all
+netty-
 nimbus-jose-jwt
 oauth2-oidc-sdk
 rmiio
@@ -339,6 +340,7 @@ if [ "${licFromWs}" = "true" ] ; then
     echo "Incorrect LICENSE in binary distribution"
     echo "Expected:" $(wc -c ${BLICENSE})
     echo "Actual:" $(wc -c ${NEW_DIR}/LICENSE)
+    result=1
   fi
 
   if ! [ "$SKIP_SRC_LICENSE" = "true" ] ; then
@@ -349,13 +351,15 @@ if [ "${licFromWs}" = "true" ] ; then
       echo "Incorrect LICENSE in source distribution"
       echo "Expected:" $(wc -c ${SLICENSE})
       echo "Actual:" $(wc -c ${NEW_SRC_DIR}/LICENSE)
+      result=1
     fi
 
     banner "Checking references in source license"
-    cat $SLICENSE | checkMissing $SLICENSE
+    cat $SLICENSE | checkMissing $SLICENSE || result=1
+
 
     banner "Checking references in binary license"
-    cat $SLICENSE $SLICENSE $BLICENSE | sort | uniq -u | checkMissing $BLICENSE
+    cat $SLICENSE $SLICENSE $BLICENSE | sort | uniq -u | checkMissing $BLICENSE || result=1
   fi
 fi
